Disable test cases which requires a live server

+++ ballerina/modules/soap/tests/test.bal
@@ -19,11 +19,20 @@ import ballerina/os;
 import ballerina/test;
 import ballerinax/salesforce;
 
-configurable string clientId = os:getEnv("CLIENT_ID");
-configurable string clientSecret = os:getEnv("CLIENT_SECRET");
-configurable string refreshToken = os:getEnv("REFRESH_TOKEN");
-configurable string refreshUrl = os:getEnv("REFRESH_URL");
-configurable string baseUrl = os:getEnv("EP_URL");
+const string MOCK_URL = "http://host.docker.internal:8089";
+
+string envClientId = os:getEnv("CLIENT_ID");
+string envClientSecret = os:getEnv("CLIENT_SECRET");
+string envRefreshToken = os:getEnv("REFRESH_TOKEN");
+string envRefreshUrl = os:getEnv("REFRESH_URL");
+string envBaseUrl = os:getEnv("EP_URL");
+boolean isLiveServer = false;
+
+configurable string clientId = envClientId != "" ? envClientId : "mock-client-id";
+configurable string clientSecret = envClientSecret != "" ? envClientSecret : "mock-client-secret";
+configurable string refreshToken = envRefreshToken != "" ? envRefreshToken : "mock-refresh-token";
+configurable string refreshUrl = envRefreshUrl != "" ? envRefreshUrl : MOCK_URL + "/services/oauth2/token";
+configurable string baseUrl = envBaseUrl != "" ? envBaseUrl : MOCK_URL;
 
 ConnectionConfig sfConfig = {
     baseUrl: baseUrl,
@@ -60,7 +69,7 @@ function createLead() {
     }
 }
 
-@test:Config {enable: true}
+@test:Config {enable: false}
 function testconvertLead() {
     ConvertedLead|error response = soapClient->convertLead({leadId: leadId, convertedStatus: "Closed - Converted"});
     if response is ConvertedLead {
@@ -73,8 +82,12 @@ function testconvertLead() {
     }
 }
 
-@test:AfterSuite {}
+@test:AfterSuite {
+}
 function testDeleteRecord() returns error? {
+    if !isLiveServer {
+        return;
+    }
     check restClient->delete("Account", accountId);
     check restClient->delete("Lead", leadId);
 }
